+++ packages/plugins/type-distributor/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added automatic ld0 creation to the accespoints. It will include all l-class lnodes except LGOS and LSVS.
 - Added automatic lln0 creation to the ldevice when it is created.
 - Allow user to reopen equipment matching panel before the first lnode assignment
+- Added validation for IedName: min-max 1-64 chars, start with letter, only alphanumeric, can not be "None"
 - `LDevice` elements now receive a `ldName` attribute set to `IedName_inst` on creation.
 - `LD0` inst and ldName now encode the access point name: `LD0_APname` / `IedName_LD0_APname`.
 ### Changed